functions/langgraph-rag/requirements.txt
functions-framework==3.*
langchain==1.0.3
langgraph==1.0.2
langchain-google-vertexai==2.1.2
google-cloud-secret-manager
pinecone
# google-cloud-storage
